add DailyActivity

+++ FitnessApi/Dto/MealItem/MealItemDTO.cs
@@ -9,7 +9,7 @@ namespace FitnessApi.Dto.MealItem
         public decimal Protein { get; set; }
         public decimal Fat { get; set; }
         public decimal Carbohydrates { get; set; }
-        public int Kcal { get; set; }
+        public decimal Kcal { get; set; }
         public string ServingSize { get; set; }
         public string DataSource { get; set; }
         public string Name { get; set; }
